Make sure sub-events cannot be cyclical (#11710)
## Summary of the problem

In https://github.com/hackclub/hcb/pull/11705 I added a mechanism for
admins to set the parent for a given event. This open up the possibility
for cycles in the event &rarr; sub-event graph which would cause
infinite loops in a few places.

## Describe your changes

Adds a validation to `Event` to ensure `parent_id` cannot be changed to
a value that would create a cycle.