Use fixed instead of absolute positioning for admin flash (#11727)
## Summary of the problem

In #11398 I switched the admin flash to a toast. In the course of
developing it we switched from it appearing at the top of the screen to
the bottom, at which point it didn't occur to me that `position:
absolute` would result in it scrolling up with the rest of the page
should the page height exceed the viewport.


https://github.com/user-attachments/assets/7e3b141b-4681-4212-b0ec-b4c9b926d513

## Describe your changes

Use fixed positioning.