Remove promotional banner + Make robotics raffle visible (#11810)
## Summary of the problem


1. We want to remove the promotional banner
2. We want the robotics 3d printer raffle to be visible to Argosy

## Describe your changes

This is kinda a semi-hot fix to get it out there. It gets what we want,
but i'm not super happy about the extra `@is_argosy` variable.

<img width="1035" height="645" alt="image"
src="https://github.com/user-attachments/assets/80ce53bb-e706-4127-bdcb-212e8ce04227"
/>

Orgs that have the promotions feature disabled within the Event plan
will now be able to see the promotions page (previously they couldn't),
but all promotions on that page will be disabled.

The only exception is Argosy, which has promotions disabled, but is
_**only**_ able to view the robotics 3d printer raffle promotion.