[Wrapped] Abstract year into the Metric class (#11830)
## Summary of the problem
Most `Metric`s are all based on data from the current or previous year.
It can be a pain to update the year in every place, and it makes it hard
to iterate on data for a new year.

## Describe your changes
This PR removes all hardcoded instances of `2024` in any metric and
instead calls `Metric.year`.