Replace session object with just IP address in StripeCardService::Create (#11858)
## Summary of the problem
For #11772 I can't use the `current_session` object as it is not defined
in the API and we are only passing in current session for the IP address
in the service.


## Describe your changes
I just changed the parameters to just take in the IP address instead of
the whole session object.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>