Fix finalized_at field failing to be set to a timestamp (#11872)
## Summary of the problem

The Rails 7 behavior was to silently set it to nil and it was never
being assigned correctly. Rails 8 will raise a
`PG::DatetimeFieldOverflow` error.

- https://github.com/rails/rails/issues/53679
- https://github.com/rails/rails/issues/54110


## Describe your changes

PR extracted from last commit of
https://github.com/hackclub/hcb/pull/10836/commits/47a94a6dd62c15ab322c4198e4f680cf1f9b0afd#r2203418417

Bringing over some additional @KiKoS0 comments

> I'm not sure if there are other occurrences of this that aren't
covered by tests; If there are it might not be a bad idea to monkey
patch it like in:
https://github.com/rails/rails/issues/53679#issuecomment-2530827971

> not really sure why we need respond_to? when status_transitions isn't
nullable:
https://docs.stripe.com/api/invoices/object?api-version=2024-06-20#invoice_object-status_transitions.
also doesn't look like the Invoice object can ever have a finalized_at
field.

Co-authored-by: Riadh <22998716+KiKoS0@users.noreply.github.com>