Don't require a signed in user for receipt status image (#11873)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
https://hackclub.slack.com/archives/GBXBFEED9/p1759940792433679
In Gmail, a broken image is displayed where it should be a button to
upload the receipt.
Originally, we thought this was due to Gmail not rendering SVGs
correctly. However, it seems that we were actually requiring a signed in
user to fetch the image, which would not be sent when viewing from any
email client, but it would exist when viewing from hcb.hackclub.com,
letter opener, or previews in development.

It's still possible that SVGs could not work in Gmail, but I'll test to
see if it works after this PR is merged and deployed.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Bypass requiring a signed in user for the receipt status action.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>