Update eligibility criteria text for premium pins (#11883)
Added eligibility note for US-based organizations.

## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->

Costs for intl shipping are too high- decision has been made to be
US-only for the time being until we can find a solution.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->

Added text that mentions it is for US-based organizations only.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>