[Subevents] Add scoped user-facing tags (#11907)
## Summary of the problem
Subevents should be able to be tagged with user-facing tags to help
filter them!

Closes #11899

## Describe your changes
Adds a new `Event::ScopedTag` model, which is the new user-facing tag
model that is scoped to a parent event. This model has a many-to-many
relationship with `Event` through `Event::ScopedTagsEvent`.

Managers of the parent event can create, destroy, and add tags to a new
subevent they are creating within the form using a combobox implemented
with Alpine, similar to the `PaymentRecipient` combobox. Tags can also
be created, deleted, and toggled from the add tag menu on each event
card on the sub-organizations page.

Sub-organizations can be filtered by tag, which will also update the
balance shown at the top of the sub-organizations page.


https://github.com/user-attachments/assets/4d853ca3-ceac-4a3c-b647-9db92285f596

**note:** I've disabled ERB lint for
`app/views/events/sub_organizations.html.erb` and
`app/views/events/_scoped_tag_option.html.erb` as that is what was done
for the other files where Alpine is used, since erb_lint doesn't like
HTML attributes that start with `@`. I'll make a PR after this to change
all `@click` attributes to use `x-on:click` to avoid needing to disable
ERB lint for the whole file. Other than issues related to the `@click`
attribute, these files pass the lints.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Gary Tou <gary@garytou.com>