[v4] Preload only the paginated event transactions (#11926)
## Summary of the problem
API times to load big orgs like HQ or HCB Ops is wayyyyyy longer than
ledger. Why? The code is preloading associations for ALL transactions,
and we're only showing 25 of them.

## Describe your changes
This moves the association preloader logic after we have paginated them,
so only preloads the 25 or whatever transactions instead of all.


<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: yodalightsabr <yoda@yodacode.xyz>