[Discord] Don't rely on HCB code being passed through `custom_id` (#11947)
## Summary of the problem

We're requiring that the HCB code exists at the time of activity sending
(which isn't always the case), and we're also relying on the HCB passed
through Discord.


## Describe your changes

By computing the HCB code from the Discord message ID, we're able to
rely less on the custom ID given by Discord. And, by removing this
dependency, we can skip passing in an HCB code when sending the public
activity notification, and we're less susceptible to race conditions
where the HCB code doesn't yet exist.