[Announcements] Unlink transparency mode and monthly announcements (#11945)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
See #11944 for context - for now, we've decided to not automatically
link transparency mode to enabling monthly announcements as it doesn't
give the event managers any chance to get context on what monthly
announcements are.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Remove the callbacks on `Event::Configuration` and `Event` that
automatically enabled monthly announcements if the event was/became
transparent.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->