[Announcements] Improve announcement explanations (#11946)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
There was some confusion about monthly announcements when the canceled
emails got sent out a couple days ago.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Adds the announcement explanation partial to the canceled email and adds
a link to the new monthly announcement blog post in the explanation
partial.

This PR should not be merged until
https://github.com/hackclub/hcb-engr/pull/108 is merged and deployed.