[Admin] Improve performance of admin email list (#11952)
## Summary of the problem

This page was getting unusable. The biggest culprit was generating a
select field with EVERY user in the database!

And then, with a bunch of iframes on the page, browser extensions were
going crazy trying to load into every iframe.


## Describe your changes

Turboframe the iframes... yes, I know... crazy. See
https://github.com/hackclub/hcb/blob/c8b67027858888deb0084ffb15a81ae4f7a30607/app/views/admin/email.html.erb
for an explanation.

Turn the User select dropdown into a simple `user_id` text field. Yeah,
it sucks now, but at least the page will load.

---------

Co-authored-by: yodalightsabr <yoda@yodacode.xyz>