[CPT Mailer] Switch to PNGs for receipt status image (#11906)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
#11873 tried to fix an issue where receipt status images did not display
in Gmail. While that PR did fix an issue affecting all email clients, it
seems that Gmail still doesn't display the image as it is an SVG.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Re-exported the receipt buttons from Figma as a PNG and modified the
controller to send the appropriate PNG from the assets directory.