[Announcements] Add job to automatically enable monthly announcements (#12018)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
Monthly announcements have been in production for a few months now, but
we haven't really rolled them out well yet, and there's been some
confusion about them for events that got them automatically enabled
without any notice.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
We've decided to only automatically enable monthly announcements for
transparent teen-led events. This PR adds a new one time job,
`OneTimeJobs::EnableMonthlyAnnouncements` to do this. It finds all
eligible events that do not already have monthly announcements enabled,
enables it, and sends a notice email that explains the feature and links
to the new monthly announcement that will be sent out on December 1st.

A parameter `exclude_ids` is added to the job to make sure we don't
automatically enable the feature for any events that we just sent the
clarification email in #11944 to - it would be weird if someone disabled
monthly announcements after receiving that email, and then we
automatically enabled it for them _again_.

To run in production after merging:
```ruby
exclude_ids = [...] # we have saved the list of IDs that we sent the clarification email to
OneTimeJobs::EnableMonthlyAnnouncements.perform_now(exclude_ids:)
```

<img width="1890" height="923" alt="image"
src="https://github.com/user-attachments/assets/2831777c-97cc-426d-996f-be855e255a47"
/>

---------

Co-authored-by: Gary Tou <gary@garytou.com>