Upgrade strong_migrations to v2 (#12071)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
After our migration to Rails 8 started (#11976), we started seeing that
`rails db:migrate` was no longer working as expected - it did not detect
new migrations to run. This was caused by our old version of the
`strong_migrations` gem, which had been held back to major version 1
because we still used Postgres 11 in development. Now that we use
Postgres 15 in development, we can safely upgrade `strong_migrations` to
the latest version.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Upgrades `strong_migrations` to version 2.5.1.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->