Disable send grant button according to card grant policy (#12152)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
Currently, the send grant button on the new card grants overview page
checks the transfer create policy for if it is disabled. However, this
policy differs from the card grant create policy because it does not
check if the event plan has card grants enabled.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Changes the disabled attribute on the send grant button to check the
card grant policy instead of the transfer create policy.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

Co-authored-by: Luke Oldenburg <87272260+Luke-Oldenburg@users.noreply.github.com>