Move contracts to a new `Contract` model (#12094)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
Currently, contracts are in an `OrganizerPosition::Contract` model and
tightly coupled to `OrganizerPositionInvite`. With the new application
flow I'm working on, we'll need contracts to be abstracted more to
support a source from either an `Event::Application` or
`OrganizerPositionInvite`, since `OrganizerPositionInvite` will not
exist yet when the applicant is signing the contract during the
application flow.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Primarily, this PR adds a new `Contract` model that is mostly the same
as `OrganizerPosition::Contract`, with a couple differences:
- Has a polymorphic relationship `Contractable` instead of a
relationship directly to `OrganizerPositionInvite`. Only
`OrganizerPositionInvite` implements this right now, but in the
applications PR `Event::Application` will also implement it.
- A `type` column instead of a `purpose` column. The `purpose` column
was added in the original implementation of
`OrganizerPosition::Contract`, but it is not currently being used. This
PR adds this back as STI, and a future PR will implement better
automation for termination contracts

All code that references `OrganizerPosition::Contract` has been changed
to reference `Contract` instead, including the relations on `User` and
`Event`. A future PR will drop the `OrganizerPosition::Contract` model
and associated table.

In the database migration that creates this new table, steps have been
added that copy over current data from the
`organizer_position_contracts` table and renames item types for versions
stored in the PaperTrail `versions` table.

I've tested this PR using my own DocuSeal account in test mode, but it
should be tested with our production DocuSeal account and a real
contract before merging into production.

---------

Co-authored-by: Gary Tou <gary@garytou.com>