Only create Contracts through Contractable (#12251)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->

https://appsignal.com/hack-club/sites/6596247683eb67648f30f807/exceptions/incidents/2394

Bug introduced in #12094 - since `Contract` and
`OrganizerPositionInvite` are now decoupled, there are some fields of
the `Contract` model that need to be set using info from the
`Contractable`. Notably, the `prefills` jsonb includes event data that
must be included in the `Contract` before creation, since it is used in
the after create callback that sends the payload to Docuseal. While we
were doing this in the `send_contract` method on
`OrganizerPositionInvite`, we were not doing this when creating a
contract using the create action in `ContractsController` (used when
sending a contract on a pre-existing invite or organizer position),
causing the above bug.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
While this could have been fixed by just grabbing the necessary data
from the `Contractable` and passing it to `Contract.new` in
`ContractsController`, it's better to maintain that contracts should
only be created using methods on the `Contractable`. This way, changes
can be easily made to what data is passed to the `Contract`, and any
extra code (such as updating the `OrganizerPosition` or Airtable can be
done in one place.

This does involve updating the `OrganizerPositionInvite` after it has
been sent and accepted - I've fixed a validation to not fail when this
happens, but it would also be possible to modify the code so that this
model is essentially "frozen" after being accepted, and only update the
`is_signee` field before accepting.