Fix earmuffed mail error and include cc and bcc recipients (#12318)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
We've been seeing a ton of
https://appsignal.com/hack-club/sites/6596247683eb67648f30f807/exceptions/incidents/2520/samples/timestamp/2025-12-04T21:24:48Z
lately, which is caused by the new earmuffed mail feature. In short, the
`after_action` method was constructing the mail object to see if it
contained recipients, but this sometimes caused errors when `mail` is
not called in the mailer and the default mail object would not be valid.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Changes earmuffed recipients to work by overriding the `mail` method
itself to remove earmuffed recipients from the to, cc, and bcc headers
of the mail object (currently, this only acts on the to header, but
comment emails only use bcc). This prevents constructing a mail object
when `mail` is never called.

---------

Co-authored-by: Gary Tou <gary@garytou.com>