Handle nil signee party on the home page (#12432)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
While we shouldn't have any sent contracts that don't have an associated
signee party, we seem to have missed some in the backfill (see
https://hcb.hackclub.com/blazer/queries/1051-sent-contracts-without-a-signee-role).
This is causing the HCB home page to 500 for anyone who has one of these
contracts.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
While we work on sorting out this data, this PR adds a check to make
sure we don't try to render any contracts that don't have a signee
party.

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->