Fix issues for HCB's contract party notification (#12422)
## Summary of the problem
<!-- Why are these changes being made? What problem does it solve? Link
any related issues to provide more details. -->
HCB Operations was getting notiifed to sign the fiscal sponsorship
agreement before other parties had signed their part. The email we sent
to them was also unclear that it was for them to sign as operations and
not as the contract signee.

## Describe your changes
<!-- Explain your thought process to the solution and provide a quick
summary of the changes. -->
Waits to notify HCB's party until all other parties have signed, and
edits the email subject and body for HCB's notification. Includes giving
them back their cookies - they deserve it 🍪

<!-- If there are any visual changes, please attach images, videos, or
gifs. -->

---------

Co-authored-by: Luke Oldenburg <87272260+Luke-Oldenburg@users.noreply.github.com>